[env](compiler)Optimize the compilation time of cast. (#61276)
Previously, many cast instantiations were done in .h files, which caused
a lot of code to be instantiated whenever other files included them.
Here we move those instantiations into .cpp files, and keep only the
code for casting a single value in the .h file.
