+++ be/src/exprs/function/cast/cast_to_float.h
@@ -96,48 +96,5 @@ public:
                                                                  input_rows_count);
     }
 };
-
-namespace CastWrapper {
-
-// max float: 3.40282e+38
-// max int128:
-// >> 0x7fffffffffffffffffffffffffffffff
-// 170141183460469231731687303715884105727
-// >>> len('170141183460469231731687303715884105727')
-// 39
-template <typename ToDataType>
-WrapperType create_float_wrapper(FunctionContext* context, const DataTypePtr& from_type) {
-    std::shared_ptr<CastToBase> cast_impl;
-
-    auto make_cast_wrapper = [&](const auto& types) -> bool {
-        using Types = std::decay_t<decltype(types)>;
-        using FromDataType = typename Types::LeftType;
-        if constexpr (type_allow_cast_to_basic_number<FromDataType>) {
-            if (context->enable_strict_mode()) {
-                cast_impl = std::make_shared<
-                        CastToImpl<CastModeType::StrictMode, FromDataType, ToDataType>>();
-            } else {
-                cast_impl = std::make_shared<
-                        CastToImpl<CastModeType::NonStrictMode, FromDataType, ToDataType>>();
-            }
-            return true;
-        } else {
-            return false;
-        }
-    };
-
-    if (!call_on_index_and_data_type<void>(from_type->get_primitive_type(), make_cast_wrapper)) {
-        return create_unsupport_wrapper(
-                fmt::format("CAST AS number not supported {}", from_type->get_name()));
-    }
-
-    return [cast_impl](FunctionContext* context, Block& block, const ColumnNumbers& arguments,
-                       uint32_t result, size_t input_rows_count,
-                       const NullMap::value_type* null_map = nullptr) {
-        return cast_impl->execute_impl(context, block, arguments, result, input_rows_count,
-                                       null_map);
-    };
-}
-} // namespace CastWrapper
 #include "common/compile_check_end.h"
 } // namespace doris
