+++ be/src/exprs/function/cast/function_cast.cpp
@@ -22,18 +22,11 @@
 #include "core/data_type/data_type_number.h" // IWYU pragma: keep
 #include "core/data_type/primitive_type.h"
 #include "exprs/function/cast/cast_to_array.h"
-#include "exprs/function/cast/cast_to_boolean.h"
-#include "exprs/function/cast/cast_to_date.h"
-#include "exprs/function/cast/cast_to_decimal.h"
-#include "exprs/function/cast/cast_to_float.h"
-#include "exprs/function/cast/cast_to_int.h"
-#include "exprs/function/cast/cast_to_ip.h"
 #include "exprs/function/cast/cast_to_jsonb.h"
 #include "exprs/function/cast/cast_to_map.h"
-#include "exprs/function/cast/cast_to_string.h"
 #include "exprs/function/cast/cast_to_struct.h"
-#include "exprs/function/cast/cast_to_timestamptz.h"
 #include "exprs/function/cast/cast_to_variant.h"
+#include "exprs/function/cast/cast_wrapper_decls.h"
 #include "exprs/function/simple_function_factory.h"
 
 namespace doris {
@@ -244,45 +237,31 @@ WrapperType prepare_impl(FunctionContext* context, const DataTypePtr& origin_fro
     case PrimitiveType::TYPE_BOOLEAN:
         return create_boolean_wrapper(context, from_type);
     case PrimitiveType::TYPE_TINYINT:
-        return create_int_wrapper<DataTypeInt8>(context, from_type);
     case PrimitiveType::TYPE_SMALLINT:
-        return create_int_wrapper<DataTypeInt16>(context, from_type);
     case PrimitiveType::TYPE_INT:
-        return create_int_wrapper<DataTypeInt32>(context, from_type);
     case PrimitiveType::TYPE_BIGINT:
-        return create_int_wrapper<DataTypeInt64>(context, from_type);
     case PrimitiveType::TYPE_LARGEINT:
-        return create_int_wrapper<DataTypeInt128>(context, from_type);
+        return create_int_wrapper(context, from_type, to_type->get_primitive_type());
     case PrimitiveType::TYPE_FLOAT:
-        return create_float_wrapper<DataTypeFloat32>(context, from_type);
     case PrimitiveType::TYPE_DOUBLE:
-        return create_float_wrapper<DataTypeFloat64>(context, from_type);
+        return create_float_wrapper(context, from_type, to_type->get_primitive_type());
     case PrimitiveType::TYPE_DATE:
-        return create_datelike_wrapper<DataTypeDate>(context, from_type);
     case PrimitiveType::TYPE_DATETIME:
-        return create_datelike_wrapper<DataTypeDateTime>(context, from_type);
     case PrimitiveType::TYPE_DATEV2:
-        return create_datelike_wrapper<DataTypeDateV2>(context, from_type);
     case PrimitiveType::TYPE_DATETIMEV2:
-        return create_datelike_wrapper<DataTypeDateTimeV2>(context, from_type);
+    case PrimitiveType::TYPE_TIMEV2:
+        return create_datelike_wrapper(context, from_type, to_type->get_primitive_type());
     case PrimitiveType::TYPE_TIMESTAMPTZ:
         return create_timestamptz_wrapper(context, from_type);
-    case PrimitiveType::TYPE_TIMEV2:
-        return create_datelike_wrapper<DataTypeTimeV2>(context, from_type);
     case PrimitiveType::TYPE_IPV4:
-        return create_ip_wrapper<DataTypeIPv4>(context, from_type);
     case PrimitiveType::TYPE_IPV6:
-        return create_ip_wrapper<DataTypeIPv6>(context, from_type);
+        return create_ip_wrapper(context, from_type, to_type->get_primitive_type());
     case PrimitiveType::TYPE_DECIMALV2:
-        return create_decimal_wrapper<DataTypeDecimalV2>(context, from_type);
     case PrimitiveType::TYPE_DECIMAL32:
-        return create_decimal_wrapper<DataTypeDecimal32>(context, from_type);
     case PrimitiveType::TYPE_DECIMAL64:
-        return create_decimal_wrapper<DataTypeDecimal64>(context, from_type);
     case PrimitiveType::TYPE_DECIMAL128I:
-        return create_decimal_wrapper<DataTypeDecimal128>(context, from_type);
     case PrimitiveType::TYPE_DECIMAL256:
-        return create_decimal_wrapper<DataTypeDecimal256>(context, from_type);
+        return create_decimal_wrapper(context, from_type, to_type->get_primitive_type());
     case PrimitiveType::TYPE_CHAR:
     case PrimitiveType::TYPE_VARCHAR:
     case PrimitiveType::TYPE_STRING:
